Functioning pixelate effect (horrible performance though), added resource loader usage when opening project via menus or from file explorer and also fixed a bug where removing entries (either offlining or them loading successfully) didn't actually remove them :/ And added Long parameter property editor slot

+++ FramePFX/Editors/DataTransfer/DataParameterLong.cs
@@ -34,9 +34,9 @@ namespace FramePFX.Editors.DataTransfer {
             this.hasRangeLimit = minValue != long.MinValue || maxValue != long.MaxValue;
         }
 
-        public double Clamp(double value) => Maths.Clamp(value, this.Minimum, this.Maximum);
+        public long Clamp(long value) => Maths.Clamp(value, this.Minimum, this.Maximum);
 
-        public bool IsValueOutOfRange(double value) => value < this.Minimum || value > this.Maximum;
+        public bool IsValueOutOfRange(long value) => value < this.Minimum || value > this.Maximum;
 
         public override void SetValue(ITransferableData owner, long value) {
             if (this.hasRangeLimit) {
